pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "access-governance-env"
version = "0.1.0"
description = "OpenEnv environment for deterministic internal access governance review."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "gradio>=5.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["access_governance_env", "access_governance_env.server", "server"]
